Move tickets this week to dashboard stats
