+++ pages/index.tsx
@@ -11,7 +11,6 @@ import TicketsTable from '@components/TicketsTable';
 import SampleDashboardStats from '@components/SampleDashbordStats';
 import useRecentlyClosedTickets from '@hooks/useRecentlyClosedTickets';
 import useMyTickets from '@hooks/useMyTickets';
-import useTicketsThisWeek from '@hooks/useTicketsThisWeek';
 
 const Dashboard: React.FC = () => {
   const { data: myTickets, isLoading: loadingMyTickets } = useMyTickets();
@@ -20,27 +19,11 @@ const Dashboard: React.FC = () => {
     data: recentlyClosedTickets,
     isLoading: loadingRecentlyClosedTickets,
   } = useRecentlyClosedTickets();
-  const ticketsThisWeek = useTicketsThisWeek();
 
   return (
     <div className="grid">
       <SampleDashboardStats />
       <div className="col-12 xl:col-6">
-        <div className="grid">
-          <div className="col-12 lg:col-12 xl:col-6 mb-2">
-            <div className="card mb-0">
-              <div className="flex justify-content-between mb-3">
-                <div>
-                  <span className="block text-500 font-medium mb-3">New Cases This Week</span>
-                  <div className="text-900 font-medium text-xl">{ticketsThisWeek.length}</div>
-                </div>
-                <div className="flex align-items-center justify-content-center bg-teal-100 border-round" style={{ width: '2.5rem', height: '2.5rem' }}>
-                  <i className="pi pi-calendar text-teal-500 text-xl" />
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
         <div className="card">
           <h5>My Cases</h5>
           <TicketsTable
