+++ src/components/SampleDashbordStats.tsx
@@ -1,6 +1,23 @@
+import useTicketsThisWeek from '@hooks/useTicketsThisWeek';
+
 export default function SampleDashboardStats() {
+  const ticketsThisWeek = useTicketsThisWeek();
+
   return (
     <div className="grid col-12">
+      <div className="col-12 lg:col-6 xl:col-3">
+        <div className="card mb-0">
+          <div className="flex justify-content-between mb-3">
+            <div>
+              <span className="block text-500 font-medium mb-3">New Cases This Week</span>
+              <div className="text-900 font-medium text-xl">{ticketsThisWeek.length}</div>
+            </div>
+            <div className="flex align-items-center justify-content-center bg-teal-100 border-round" style={{ width: '2.5rem', height: '2.5rem' }}>
+              <i className="pi pi-calendar text-teal-500 text-xl" />
+            </div>
+          </div>
+        </div>
+      </div>
       <div className="col-12 lg:col-6 xl:col-3">
         <div className="card mb-0">
           <div className="flex justify-content-between mb-3">
@@ -44,21 +61,6 @@ export default function SampleDashboardStats() {
           </div>
           <span className="text-green-500 font-medium">520 </span>
           <span className="text-500">newly registered</span>
-        </div>
-      </div>
-      <div className="col-12 lg:col-6 xl:col-3">
-        <div className="card mb-0">
-          <div className="flex justify-content-between mb-3">
-            <div>
-              <span className="block text-500 font-medium mb-3">Comments</span>
-              <div className="text-900 font-medium text-xl">152 Unread</div>
-            </div>
-            <div className="flex align-items-center justify-content-center bg-purple-100 border-round" style={{ width: '2.5rem', height: '2.5rem' }}>
-              <i className="pi pi-comment text-purple-500 text-xl" />
-            </div>
-          </div>
-          <span className="text-green-500 font-medium">85 </span>
-          <span className="text-500">responded</span>
         </div>
       </div>
     </div>
